доделал ДЗ#3
