+++ src/components/MessageForm.js
@@ -1,9 +1,27 @@
-import './css/style.css';
+import '../css/style.css';
 import React, { useEffect, useState, useRef } from 'react';
-import { TextField } from "@material-ui/core";
+import { TextField, withStyles } from "@material-ui/core";
 import MuiButtons from './MuiButtons';
 
+
 const userName = 'Вадим';
+
+const MessageInput = withStyles({
+	root: {
+		'& input ': {
+			width: "500px",
+			color: "#FFFFFF",
+		},
+		'& input + fieldset': {
+			borderColor: '#b1b1b1',
+			borderWidth: 2,
+		},
+		'& label': {
+			color: "#FFFFFF"
+		},
+	},
+})(TextField);
+
 
 export default function MessageForm(props) {
 	const [value, setValue] = useState("");
@@ -30,13 +48,12 @@ export default function MessageForm(props) {
 
 	return (
 		<form className="message-form" action="#" onSubmit={handleSubmit}>
-			<TextField
+			<MessageInput
 				autoFocus
-				className="message-form__input"
+				className='message-form__input'
 				id="outlined-basic"
-				label="Outlined"
+				label="Напишите сообщение..."
 				onChange={handleChange}
-				placeholder="Напишите сообщение..."
 				ref={inputRef}
 				required
 				type="text"
