+++ src/components/MuiButtons.js
@@ -8,6 +8,7 @@ import clsx from 'clsx';
 const useStyles = makeStyles((theme) => ({
 	button: {
 		margin: theme.spacing(1),
+		color: "#FFFFFF"
 	},
 }));
