+++ src/css/style.css
@@ -177,15 +177,6 @@ code {
   margin: 30px 0 0 0;
 }
 
-.message-form__input {
-  width: 500px;
-  color: white;
-}
-
-.message-form__input::placeholder {
-  color: #B1B1B1;
-}
-
 .message-list {
   display: flex;
   flex-direction: column;
@@ -228,7 +219,7 @@ code {
 
 .message__date {
   margin: 0 0 0 5px;
-  color: #B1B1B1;
+  color: #b1b1b1;
   font-size: 10px;
   line-height: normal;
 }
