+++ src/scss/_Chat.scss
@@ -34,7 +34,7 @@
 
     &__date {
         margin: 0 0 0 5px;
-        color: #B1B1B1;
+        color: #b1b1b1;
         font-size: 10px;
         line-height: normal;
     }
