+++ src/scss/_MessageForm.scss
@@ -5,23 +5,4 @@
 	min-height: 50px;
 	position: relative;
 	margin: 30px 0 0 0;
-
-	&__input {
-		width: 500px;
-		// box-sizing: border-box;
-		// min-width: 400px;
-		// min-height: 50px;
-		// padding: 10px 80px 10px 20px;
-		// background-color: transparent;
-		// border: 1px solid #E1E1E1;
-		// border-radius: 3px;
-		// outline: none;
-		// font-family: inherit;
-		// @include font (14px, 17px, 400);
-		color: white;
-	}
-
-	&__input::placeholder {
-		color: #B1B1B1;
-	}
 }
